repo init
I've initialised a new repo that's gonna be a more efficient zig rewrite
of my old thoughts-cli. The file-format for this database will be
flexable enough so that it's future proof as long as the future programs
stay true to the essence of this program.

I've initialised all the necessary files, and also added datetime as
a dependency as that's the main/only depenency that I *need*.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>