stack: added functions for a file-based stack
I've implemented simple file-based stack functions that either push
pop or 'read' from the stack. There are only two state variables used
in the stack, the stack pointer, and also a read pointer for reading
from start to end without popping.

The format of each of the stack entries are that they are the contents
wrapped with the length of that very content at both the start and end
as, the one at the start aids in reading from start to finish, while
the one at the end aids in popping entries off the stack quickly and
efficently without having to traverse the ENTIRE stack.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>