stack items: added features and packing & unpacking
I've added code for dealing with stack items- the acutal items that will
end up being stored on the stack. Each stack item has the layout of the
feature bitfield, the feature data after it
(dynamically sized based upon what features are enabled) and also the
actual contents after that.

I've chosen to split the features into a features bitfield, which is
encoded and stored on the stack, and a features struct that is only
stored in memory and acts as an more ergonomic 'frontend' to the
features. This is so it's much easier and concise to use and read the
data from the features in the code as it's all the decoding is done in
one go. Also the state from the offset, as the size and therefore
indexes of each of the features' data is also dynamic and hard to track
if not done in one function call.

I've also changed a bunch of code to use `@sizeOf` instead of
`@bitSizeOf / 8` as the former is essentially the same as the latter
except for non-full byte values, which all of the data dealed with is.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>