added cli, wipe command & made id mandatory
I've written a basic cli with a help message for oats.
I've also implemented the simplest command specified by the help message
which is the wipe command, where I just write the required major version
and also the initial stack pointer to the oats database file. To do this
I've also written some basic code to get the path of that file.

I've also added some constants to standardize the locations of certain
values in the database in the stack file to hopefully reduce the chances
of writing to the wrong locations.

I've also made ids mandatory as I've realized that it would borderline
impossible to combine two of these oat databases together through import
without the use of an id system, so I've re-written the spec for oats
to have the id as the very first value stored in each stack item and
updated the code to reflect such.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>