added the 'push' command & small bugfix
I've added and implemented the push command for the cli. It takes one
argument, the string value to push and then pushes it to the database,
updating the stack pointer and everything. It ensures that the database
exists and also that it's of the correct major version.

I've also added a quick seek to the push function so that it actually
writes to the location that the stack pointer is pointing to.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>