magic numbers: added a magic sequence
I've added a magic sequence to the start of the oats database file to
be able to better identify and double-check the 'validity' of the
database, especially important for when I import and combine databases
in the future.

I've also updated the locations specified in the stack file to depend
on each other sequentially so I don't have to update them individually.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>