updated the push command to push valid stack items, timestamp & bugfixes
I'd forgotten that the stack should only consist of stack items and
therefore the push command I'd implemented earlier was incomplete as it
just pushed the raw data onto the stack instead of in the stack-item
format with the required metadata.

I've also made the push command include the timestamp with the contents
and also changed the timestamp field to actually be an i64, I've also
renamed the date feature to the more correct name of 'timestamp'.

I've updated the push command to construct and pack a stack item, and
I've also discovered some bugs that were introduced when I moved the id
into it's own mandatory position instead of in the features struct.

Also note that I've bitcast the timestamp to a u64 for the id on purpose
as the id just has to be unique, and by bitcasting I've been able to
retain all the same information represented.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>