replaced copyForwards with @memcpy
I've recently discovered that @memcpy is the actual correct way to copy
memory for my purposes. In that, the data does not overlap and I need
it to be fast.

But also, by replacing copyForwards with @memcpy which requires the
source and destination slices to have the same length, I've discovered
a bug in the way that items are packed where there is unnecessary
wasted space that gets allocated and also written to the stack due
to miscalculations within the featuresSize function where I got the size
of an optional instead of the actual type (optional's include the
overhead of a tagged union, 8bits).

I've also updated the code to more explicity and strictly adhere to the
same length rule.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>