changed to use proper readers and writers and fs
I've removed my own hand-rolled readInt and writeInt functions as I've
recently realized that the reader and writers provided by zig actually
have the functions baked in.

I've also found out the reason why the reader and writers take in a
file handle instead of a pointer to a file handle, as that handle is
just a wrapper over the system provided fd (file descriptor) so the
actual state is handled by the OS instead, and therefore I should only
pass around the file and not pointers to it, and that also means, if I
seek with a file, the same state is applied to it's subsequent readers
and writers due to the shared file descriptor.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>