added the pop command & normal format and bugfixes
I've added the pop command which can now take a specified amount of
stack items to pop, otherwise defaulting to 1, as that extends the
functionality and practicality with little work.

I've also added 'normal' formatting for the features of the stack-item
where it gets written to a stream as that avoids allocation and is
simpler as I will only use it for streams. The normal format is for cli
use and simple viewing, while in future, there will be a markdown format
aswell for outputting to files.

I've also fixed some bugs to do with popping and also unpacking the
stack items, mostly due to either using the wrong variable name and also
getting the size of a feature's optional.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>