added heads & tails and fixed formatting issues
I've added the heads and tails commands for viewing the contents of the
oats database.
The tails command actually works the exact same as the pop command,
except it doesn't write any changes to the database
(to the stack pointer specifically) so it doesn't actually pop the
values off the stack but just reads it.
The heads command is the same as the tails command, except it uses a
read_ptr instead of a stack_ptr and starts at the start of the database
and reads till the end.

I've also fixed some formatting issues where the iso date's length
was assumed to be an incorrect length instead of the actual length (25)
in two different places. One assumed it was 26, the other 32.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>