added count command
I've added a really simple count command to quickly count how many
stack items are actually stored in the stack. Now I could do this by
having a variable that keeps track of this, but that's just another
value that needs to be updated and kept valid which ain't worth it
for a seldom done action.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>