oats session CUSTOM USER INPUT
It's taken a VERY long time, but I've been able to implement my own
user input system that keeps track of the text and also wraps it when
you reach the end; similar to a single-line text editor. I'd chosen to
do this as it would make it console independant, it would give me more
fine-grained control over the UX and also it would perhaps be fun to
implement by myself (it was not).

The 'session' is implemented through entering raw mode (through linking
to libC and using the termion system header), reading a line with
wrapping and then just printing it to the console (for now, the actual
pushing will be implemented later).

The line input is done through reading a character from stdin in raw
mode, checking if it's an escape sequence (for arrows and other future
supported keys), if so, then it would handle it in a special way, check
for other special keys (like enter and backspace) handle them also, and
if it's just a normal character, then it would instead just insert into
an arraylist based upon a cursor that gets moved by the arrow keys,
backspace and by typing.

How line wrapping works, I'm not quite sure myself as I've just spammed
different possible solutions to my issues (text writing to the wrong
place, cursor moving to the wrong place text not wrapping new lines not
getting written etc).
However, a few notable things for the design is that upon every single
update, whether it be inserting or deleting a key, all the text after
the cursor is re-written to the screen so that any deletes or inserts
are correctly presented.
Also, with line wrapping, there is a free_lines variable that gets
mutated, this is so, due to the lines getting re-printed every single
update, that unnecessary newlines aren't created every update and
instead it will just recycle any previously written newlines if there
are any (free_lines stores the amount of newlines that aren't being used
and have already been written).
% is to calculate how many characters off the value is from the end of
the coloumn.

Also, plans for a vim-like 'command system' have also been included,
where, if the first character entered is ':' then it would enter a
special state that can only be exited upon the entering of `esc`.
Commands will not have line wrapping, nor support for arrow keys.

And finally, in future, through calculating the lines, I should be able
to add suport for up and down arrows too.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>