session: added support for up & down arrows
I've added support for up and down arrows for moving around the text.
This is done through calculating an 'estimated' position by subtracting
or adding a console line to the cursor and then checking if that
position is valid, if not, then it would snap to the closest valid
target.

I've also fixed a bug with 'free_lines' by removing a previous quickfix
that was introduced to make the session work at all (previous commit)
but instead didn't work and actually introduces a new problem that is
now fixed by it's removal.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>