oats session: added actual push functionality
I've finally 'finished' the basic functionality of the oats session, in which every line is pushed to the database as it's own stack item.
I've chosen to skip empty lines, as they're empty and if you wanted
a spacer you can simply write ' '.

The session actually reads the stack pointer at push time, this is so
that if any other oat sessions or instances modify the database in any
way, the session will have the most recent version of the stack pointer
and won't write to the wrong section and corrupt the database.

I've also added a formal cli entry instead of the debug/testing setup I
had before.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>