oats session: minor quality of life improvements
There's a slight bug with the backspace where if the removed character
is the first character, the character isn't overwritten in stdout. I've
fixed this issue by just adding a dirty if clause in the backspace
handling code.

There's another bug where the next read line, instead of respecting the
free-lines and lines after the cursor of the previously read line, just
overwrites to be the line after the cursor (due to the \n). This has
now been changed through adding some 'cleanup' code to the end of the
line read, where now it skips to after the read line while respecting
it's free-lines and lines after the cursor so that it's trivial for
the next lines to be read.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>