v0.1.0: added exporting to the markdown format
I've implemented the ability to export to format
in the cli. Note that, due to limitations in the
datetime library that I use, it is unable to
detect the local timezone so the user must provide
one or else new york is taken as the default.

How the contents are converted to markdown is that
every item from the database is treated as an
element in a list in markdown. And once two items
are of different days, the full date header is
written. And if, two items's timestamps are
different by at least 8 minutes, then the smaller
time header will be written.

Some small design notes to add:
  * I've had to manually implement conversion
    between 24 hour and 12 hour time.
  * Writing is buffered to improve performance.

This is also the v0.1.0 release.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>