added sort & import commands
I've implemented the sort and import commands. For the sort command,
it's pretty simple, read all of the items into an array, get their ids
then sort.

For the import command, I've changed it to read from a file instead of
stdin, as seeking is not allowed, then use a hashmap to prevent
duplicates and then write the unsorted contents to the database
(user should sort afterwards).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>