v0.3.0: added commands to sessions
I've added a readCommands function that reads commands instead of normal
lines. It's a special mode that can be entered by hitting ':' at the
start of the line, and can be exited by either backspace til the end,
ESC, CTRL+C or CTRL+D and will disappear as if it weren't there.

So far I've added the basic necessary commands that I need.

I've also added some colour codes to the help message of the session
so it looks nicer.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>