v0.4.1: +lite metadata and optimized imports & help
I've added a lighter version of metadata for reading from the database
as I am now planning to add operations in which the entirety of the
database's stack item metadatas (features mainly) would have to be
loaded into memory at once (such as sorting, and soon exporting to md),
which would be incredibly memory intensive and inefficient as previously
to load a stack item's metadata, you would have to load it's contents
aswell. And though previously this was fine, I'm now planning to add
the ability to include images, which would take up SIGNIFICANTLY more
storage.

How the lighter metadata works, is that, when you initially read the
stack item into memory and decode it, you simply load the metadata,
alongside a 'pointer'/index to the raw stack item the metadata
references, from which, the location of the contents can be deduced
and the contents itself can be loaded into memory, only when needed.
This means that the contents can be loaded sequentially (not all at once)
whilst the metadata for the contents can be all loaded in.

I've also optimized the import function, as previously the program would
load the entire contents of the database into a hashset and then insert
the additional imported entries into the same hashset to remove
duplicates before copying the entire contents back into the database
from memory, leaving an unsorted mess. This would no longer work with
the introduction of lite metadata, as this required loading the entire
database into memory, which may now be in the GiBs (once images are
added).

The new solution, is to only grab the ids of the items in the database,
taking up significantly less memory, and storing them into an arraylist
which can then be quiried to check if a stack item exists within the
database. This assumes that the pre-existing database is valid and free
from duplicates (which it should be). It then, after sorting the id list,
iterates through all the items of the database to import, and checks if
the item exists in the database or not through binary search, if so
it's simply pushed to the database stack. This is superior to the
previous approach as it avoids rewriting/copying the entire database
eliminating the chance of data-loss, while also keeping most of the
items in the database sorted; making the subsequent sort faster.

I've also optimized the help message by just, not formatting it at
runtime.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>