v0.4.3: bugfixes: fixed various oats session bugs
The first bugfix to the oats session, was an integer overflow when
cancelling a line when there are free lines available. This was due to
how I was clearing the line as I would delete the line's contents before
cleaning up the line, so I wouldn't have the line length information
to be able to properly clean up the line.

The second bugfix is with how the cursor returns to it's starting
position. This can be known as the starting location of the printing
will always be the location of the (cursor-1) % the coloumn width, and
the writing will always write until the end of the line. So then you can
calculate how far up to return and how far left or right to return
aswell.