v0.4.4: changed line clearing behaviour
I've changed the line clearing behaviour to, instead of jumping to the
end of the line (like when enter is hit) it instead wipes the entire
contents of the line and returns to the line's starting position- as if
the line had not been written in the first place. I've chosen to do this
to make it unambiguous for what thoughts/notes have been actually
written to the oats database. It also just feels nicer.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>