v0.5.0: +session ids & cmd & markdown export revamp
The main purpose of this feature is so that I can have multiple oats
sessions writing at the same time without them conflicting with each
other. As, each oats session will now have it's own id, which will
then be used to group them together when exporting to markdown. So
regardless of the true order of the notes, the notes that belong to the
same oats session will always appear together.

Previously, if I had two things to write down at once, in two different
oats sessions, I would have to finish writing one, and then try to
remember what I wanted to write down for the second. Now I can write
down both before I forget.

I've updated the normal format to include the new session_id field and
I've also found a smarter way to calculate the max string length that a
number can have, instead of printing it, I just find the log10 of it +1.

I've chosen to make previous notes' metadata valid, by making it so that
notes from non-oats sessions (using the cli push command) will have a
session_id of null, in that, it won't have one and will be grouped
together with previous null session_id thoughts.

I've also rewritten how markdown exporting works, due to the nature of
needing to group together notes of the same session id, with no
predetermined order to things. So, this is not even close to the most
efficient approach, but it works well enough for a seldom used
operation. How the markdown exporting now works is that it iterates
through the metadata of all the items in the oats database, and then
appends them to an arraylist that's stored in a hashmap that can be
indexed by the session_id. Then I collect the ids into a list, sort them
and iterate though each of them, printing a separator and then the
contents of the collection, which can be even further sub-separated by
headers if the timestamps are different by more than 8min while bigger
changes in time, like the day, depend on the last note regardless of
collection. For notes without session_id, they are grouped together with
the last note without a session_id, granted there isn't any notes *with*
session ids separating them. Also, if the session is fewer than 4 notes
then the collection is combined with the previous collection for
aesthetic reasons.

I've also extended the session command in the cli to include an optional
session parameter and also added a :session command to the TUI which
also takes in an optional session id (otherwise taking the current
timestamp). This is so I can continue writing in a session if I
accidentally exit it.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>