v0.5.1: added windows support & rehashed termios
I've rewritten most of the code that I've written for entering raw mode
and also getting the window width (that previously depended on the c
library termios) to both be platform independant and also use the zig
std library's bindings instead.

This allows for native support for windows through conditional
compilation, yay.

Note, I chose appdata as the home for oats for no particular reason
(on windows). I also chose to create a switch type that was either the
linux console mode or the windows console mode for convenience and so
my previously written code doesn't just blow up.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>