v0.6: image support, refined markdown exporting
This is a very large commit, many details will probably be omissed.
I'll try address all the design decisions that I've made.

The main feature introduced in this commit, ofc, is the added support
for images. There has also been some major changes to how markdown
exporting works, to both polish it and also fix some bugs.

Some notable choices made about the actual pushing of the image binary,
is that, while the flag is a pretty simple and standard `is_image` flag,
the actual feature's contents is dynamically sized and changes for each
different image, as the flag itself is is_image, but the data stored in
the features section of the oats item for this flag is the filename of
the image file. This is so, I can keep some sort of identification of
the image alongside the actual image contents, and also so I can
preserve the extension of the image and delegate the responsiblity of
the image format to the markdown displayer. How the filename is stored
is the string's length is stored alongside the actual filename contents
(much like a slice), except the `usize` is converted into a platform
independant `u16` as the filename should NOT be long enough to warrent
anything above a u16.

I've also added the push-img cli command and the :img session command
to oats.
The overall syntax for both of them is pretty similar, with the command
followed by paths to images.
The filename itself is extracted from the image paths by simply
splitting by the path symbol '/' though this might not fare well in a
windows system (we'll see). And also, this has also exposed a flaw in
oats session commands, in that, "quotes" aren't correctly parsed, as
they are treated as normal characters and spaces still split them
(I'll fix later).

I've also updated the help message of the CLI to actually reflect the
current state of the commands available. ? denotes an optional argument
that has a default value, and * denotes a variable but larger than one
amount of arguments.

I've chosen against using an image_col_id as I'd initially planned as
it's pretty redundant, as the image collections are going to be next
to each other anyways, and it'd also be incredibly complex to implement
(without major code refactoring) as it would allow for images of the
same image collection to be sprinkled across the oats item collection
and have to be collected and sorted. Simply having the consecutive
images be grouped together is a much simpler solution conceptually
and to implement.

I've also updated the normal format to support images, in that, now
instead of printing out the contents of the stack items blindly
(which for images would spew binary data into the console). It will now
check if it's an image, and if so, it'll instead output the filename
of the image and have a # separator instead of | to better differentiate
between text oat itmes and image oat items.

I've moved markdown headers into their own function as they are the same
regardless of if the oats item is an image or a piece of text. I've also
removed the 4 collection limit as regardless of size, you shouldn't
be combining the thoughts from multiple sessions together as they
don't make readable sense.

Exporting to markdown now takes in an option media_path which is used
to store the media files (images for now) of the oats database that will
be referenced by the exported markdown (relatively cuz I'm lazy). I've
chosen to make this option optional as sometimes you don't want to
include the images in the markdown
(they get discarded if no path provided).

Markdown headers now also include a newline before them, so they are
more complient with the common-mark spec.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>