v0.7: is_mobile flag & markdown export tweak
I've added a mobile flag for when this program is compiled to mobile
devices (`-Dmobile=true`) as it would be nice to know which notes were
written on my phone, and which weren't.

I use a `?void` for the feature as the feature itself is just a flag and
doesn't store any additional metadata, however it has to be an optional
type due to some comptime code I've written.

I've also written a slight tweak to the markdown exportation so that now
new day headers are emphasised when they're from new collections. This
is to be more consistent with the time headers.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>