v0.7.1: oats session: vim-like ESC+A & ESC+I
I've added vim-like custom escape codes so that when ESC+A is hit, the
cursor goes to the end of the line (like in vim) and when ESC+I is hit,
the cursor snaps to the start of the line. I've chosen to do this as
it would feel nice for my muscle memory. This was also discovered to
be possible after I realized that the escape codes were just sequences
of a standard, and that you can define your own.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>