void oats items: support for both formats
I've added basic support for 'void' oats items, which are oats items that
have beens stubbed/trimmed, in that, the metadata is stored, and the
database knows of the oat item's existence, but the actual contents of
that oats item isn't. This allows for lighter versions of the database
to be written to and then later combined with the main database for
final exports. This will be useful if the database gets too large to
sync easily, especially with the introduction of pictures and perhaps
videos.

The void oat items work simply with a is_void tag, that makes it so
that when exporting to normal format or markdown, instead of reading
it's contents like usual, it instead just outputs that it has been
trimmed (as we know the contents will be empty).

So far the void oats items aren't generatable, though this is to change
once I add the 'filter' & 'trim' commands.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>