sort & import: added support for void item imports
I've changed the import command to also import non-void items even if
their void counterparts are already present in the database, as this
command should favor non-void items. This does introduce some
duplication of items, but that should be solved by the sort command.

For the sort command, now it first collects and sorts all of the
non-void oats items and then later iterates through the void items, and
then inserts them one by one, if they don't already exist. So, in the
case of a duplication of a void & non-void item, the non-void item will
be collected into the final database, whilst the void counterpart is
discarded as it already exists in the database.

I've also done utter comptime war crimes to the binary search function.
It's horribly inefficient at comptime, but whatever, it works, and it
doesn't affect runtime speeds *too* much.

I've also changed the import command to store the entire item instead of
just the id when sorting and checking for duplicates as now we need to
also know if the pre-existing item is void or not.

Also, note the binary search function now returns the 'found' location
regardless of if it actually finds anything or not. And I'm heavily
assuming that the lower bound will be where the item should be inserted
to maintain order, as the higher bound should be where it should be
inserted
         [too_low, too_high]
                  ^ insert here (higher bound)
but once it figures out that the item doesn't exist, the lower bound
is one higher than the higher bound.
         [too_high, too_low]
                    ^ insert here

Signed-off-by: Ethan Ma <kalscium@protonmail.com>