trim cmd and unpack, md export & tmp path bugfixes
I've added the trim command to the cli.
What this command does, is it takes in the attributes to trim, and the
output database path and then iterates through all the items of the oats
database and only copies over oats items that don't contain the
aformentioned attributes/features, the items that do contain them
however, are stubbed. The checking attributes part is automated by some
comptime code I've written, as I don't want to manually write the if
statements for each and every value.

Also in the unpacking of the item, I forgot to add support for unpacking the
void feature and that's why normal & markdown formats didn't export
properly.

Also I've added an optional 's' so the trimmed item markdown message is
always grammatically correct.

The oats tmp path is also now based off of the main oats path so they're
closer together and in the same directory.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>