oats database import bugfix
There was a bug where, even after import de-duplication, there were a
bunch of duplicate oats items in the database.
And, after careful investigation, I've found the cause of the issue.
Ambiguous wording.
I'd said the code did one thing, but in fact did the complete opposite,
as the continue, didn't import but in fact discarded. So, instead of
importing only the missing items, it imported only the ones that it
already had.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>