added the filter command
I've added the filter command, which is just a carbon copy of the
trim command but instead of stubbing when an attribute is found, it
stubs when a specified attribute is *not* found.

I've chosen against some sort of unifying function for this
functionality as it just doesn't make sense on it's own.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>