v0.8.1: normal, trim & oats session: minor changes
This release of oats has a few minor changes & bugfixes to the following:
  * displaying void/stubbed oats items in the normal format
  * added the 'everything' attribute option to the trim command
  * bugfix to the cleanup jump

For the normal format, we want to distinguish between the normal oats
items and the void ones even more. So, like with images, we change the
separator to something unique. I've chosen `?` to be the separator, as
it looks similar to zig optionals.

For the trim command, I've added an extra condition that checks if one
of the provided attributes is 'everything', in which case, like the name
suggests, it trims everything and outputs a database with only void
items (the metadata).

For the cleanup jump. I've noticed that when the text spanned the entire
line, and you hit enter, instead of just taking the next line, it
instead jumped after the wrapping line which is not what I want.
  |=>> wwwwwwwwwwwwwwwwwwwwwwww|
  |...                         |
  |=>> jumps here (not wanted) |
And after removing that `-1` from the jump calculation, it seemed to
solve my problem.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>