v0.8.2: finally fixed the import duplication issue
I've finally figured out the cause and fixed the import duplication
issue. It was where, after importing a database with extra oats, it
would sometimes not correctly discard the duplicates, and would keep
duplicates of the same oats item.

This was due to the binary search algorithm I was using. Not that there
were any issues with it, just that it requires an ordered list of items,
and, the way in which I added non-duplicated items to the search-list
was that I just appended them, which would mess up the binary search if
it was out of order.

I've solved this by simply inserting into the correct location
(found by the binary search).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>