added initial_text to readLine & wrapLine bugfixes
I've added the ability to include some initial text to the readLine
function that is appended to the line, printed to the screen and jumped
to the end of (by the cursor). Whilst doing this, a bug was discovered
in the wrapLine function.

In many places in the wrapLine function, the cursor is subtracted by one
so that the offsets and such are correct, but, if the cursor is at the
start of the line then it crashes by an integer underflow. This wasn't
noticed previously, as, at the start of the line, the text would be
printed *after* the key was appended, so the cursor would be at least
position 1. I've solved this by simply doing a saturating subtraction
so instead of crashing, it just clamps to zero.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>