added the ammend command to oats session
I've implemented the ammend command for oats sessions.
Everything should be pretty self-explanitory and there's not much that's
peculiar about this implmentation.

I've chosen to generate a new oats item id when ammended despite keeping
the rest of the metadata the same, as oats should NEVER have to deal
with two separate versions of the same oats item.
So, by generating a new oats item id, the item is considered as a new
separate item (with the last version dropped).
This is so, if importing another oats db with the previous version of
the ammended item, there is no ambiguity of which version survives, and
garrentees no data-loss.

Also, due to the possibility of ammends recursing
(run the ammend command within the ammend), the error union is no longer
able to be inferred, and instead I've decided to just use an anyerror
(any error used in the program is possible) as that's the simplest
solution.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>