combined clearing lines & run commands as cancels
I've combined the behaviour of clearing lines and also run commands as
'canceling' the line. As, previously, when a command was run, it would
return an empty line instead of a line cancel
(which causes problems whilst ammending).

Also, the behaviour for clearing the line is the exact same both run
commands and also cleared line so this also removes some complexity
to the state-machine.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>