oats session cleanup jump & readCommand bugfixes
Recently I've realized that line wrapping whilst ammending longer lines
(lines that spanned more than one line) had broken wrapping prompts.
This was due to me using the `wrap_prompt` variable which, when traced
is actually the wrap prompt of the readCommand, which is 6 characters
long, as apposed to the initial print prompt lenght of 4.

I had also discovered another bug with the ammend, as cancelling the
ammend would cause the cleanup jump to jump and clear one line too high.
After some investigation of the cleanup jump for both readLine and
readCommand, I've found a bunch of illogical & mislabeled code
(code that does one thing, whilst the comment says it does another).
I've rewritten the cleanup jump code for both readLine and readCommand
to actually make sense and work properly as it should.

I've also cleaned up the readCommand function of remenants of the
readLine function and also simplified it, as it doesn't need to return
it's internal line.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>