release v0.9.1: added basic tab support
Frontend changes made:
  * added basic support for tabs (2 indent)

I've added basic support 2 space tabs.
I've chosen this tab size, as it makes the most sense for a markdown
file.
So far, tabbing past the end of the coloumn (line wrapping) crashes the
program, and I don't have the care to solve it, as it would
realistically never happen with tabs, and also, the oats session code
is quickly becoming legacy, and will be rewritten in future anyways.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>