added `is_file` to the oats item features
I've added support for files in the oats item features by essentially
just copy-pasting the code already written for image filenames.
This is in anticipation of the file command which will allow me to push
generic files of any type to the oats database.
Due to backwards compatibility (as I cannot just modify the features of
the images to now support the unified is_file feature) image_filename
will remain as it's own separate feature to filename, and therefore
must also be mutually exclusive, as an oats item should not have two
filenames.
That does mean that an image cannot be a file in the logic of the
item, but you know what I mean.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>