added support for exporting to markdown & normal
I've added pretty simple support for exporting to markdown and also
normal mode. Again, through copy-pasting the pre-existing image code.

The only notable choice made is the markdown format, I've done it in
HTML as it allows or clickable linkes to files, and I've also included
that path next to it, in case the file can't be opened in a browser or
default app.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>