is_vid: added basic support in the item
I've added basic support for the is_vid flag by simply including it in
the flags. The plan for this, is that, the videos included can either
have a filename, or not, if so, the filename will be included when
exporting to markdown.

The plan is, is that the only way to include a video without a filename
is through the cli and stdin.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>