videos: +markdown & normal format support & bugfix
I've added support for exporting oats video items to either markdown or
normal format. It basically follows one to one with how an image is
exported, except for minor chanegs to suit the video, such as the
HTML being generated differently.

Also, this isn't mentioned before but, when reading from stdin, the
file-size limit of the video is set to 4GiB, this, I believe is a
reasonable comprimise as you probably shouldn't be storing videos
longer than that in the oats db.

For the generated HTML, the video is assumed to be of type mp4, as
that's the most common kind of video.
Perhaps in a future commit,
(before the release as then backwards compatibility would not allow it)
I could just add a (enum) value to the video flag that determines the kind
of video it is, this would probably be detected from either the file
format or the magic number.

Also, I've chosen for videos that do not have filenames, to simply make
the video file be it's id + .mp4 as that's a unique and fair comprimise.
(NOTE THE .MP4 MAY BE CHANGED IN FUTURE (before update))

I also noticed a bug where the binary data separator wasn't aligned with
the text separator. This was due to me forgetting to add the extra space
at the start when printing the binary separator.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>